+++ cjs/native/MoneyRecive.js
@@ -1,0 +1,8 @@
+// Archivo generado por scripts/build.mjs. No editar a mano.
+const ____core_createIcon_native_js = require('../core/createIcon.native.js')
+const { createIcon } = ____core_createIcon_native_js
+const ____data_MoneyRecive_js = require('../data/MoneyRecive.js')
+const { MoneyRecive: shape } = ____data_MoneyRecive_js
+const MoneyRecive = /*#__PURE__*/ createIcon("MoneyRecive", shape)
+
+exports.MoneyRecive = MoneyRecive