+++ cjs/native/MoneySend.js
@@ -1,0 +1,8 @@
+// Archivo generado por scripts/build.mjs. No editar a mano.
+const ____core_createIcon_native_js = require('../core/createIcon.native.js')
+const { createIcon } = ____core_createIcon_native_js
+const ____data_MoneySend_js = require('../data/MoneySend.js')
+const { MoneySend: shape } = ____data_MoneySend_js
+const MoneySend = /*#__PURE__*/ createIcon("MoneySend", shape)
+
+exports.MoneySend = MoneySend