+++ cjs/native/Setting.js
@@ -1,0 +1,8 @@
+// Archivo generado por scripts/build.mjs. No editar a mano.
+const ____core_createIcon_native_js = require('../core/createIcon.native.js')
+const { createIcon } = ____core_createIcon_native_js
+const ____data_Setting_js = require('../data/Setting.js')
+const { Setting: shape } = ____data_Setting_js
+const Setting = /*#__PURE__*/ createIcon("Setting", shape)
+
+exports.Setting = Setting