+++ cjs/web/EmptyWallet.js
@@ -1,0 +1,8 @@
+// Archivo generado por scripts/build.mjs. No editar a mano.
+const ____core_createIcon_web_js = require('../core/createIcon.web.js')
+const { createIcon } = ____core_createIcon_web_js
+const ____data_EmptyWallet_js = require('../data/EmptyWallet.js')
+const { EmptyWallet: shape } = ____data_EmptyWallet_js
+const EmptyWallet = /*#__PURE__*/ createIcon("EmptyWallet", shape)
+
+exports.EmptyWallet = EmptyWallet