+++ cjs/web/MoneyRecive.js
@@ -1,0 +1,8 @@
+// Archivo generado por scripts/build.mjs. No editar a mano.
+const ____core_createIcon_web_js = require('../core/createIcon.web.js')
+const { createIcon } = ____core_createIcon_web_js
+const ____data_MoneyRecive_js = require('../data/MoneyRecive.js')
+const { MoneyRecive: shape } = ____data_MoneyRecive_js
+const MoneyRecive = /*#__PURE__*/ createIcon("MoneyRecive", shape)
+
+exports.MoneyRecive = MoneyRecive