+++ cjs/web/MoneySend.js
@@ -1,0 +1,8 @@
+// Archivo generado por scripts/build.mjs. No editar a mano.
+const ____core_createIcon_web_js = require('../core/createIcon.web.js')
+const { createIcon } = ____core_createIcon_web_js
+const ____data_MoneySend_js = require('../data/MoneySend.js')
+const { MoneySend: shape } = ____data_MoneySend_js
+const MoneySend = /*#__PURE__*/ createIcon("MoneySend", shape)
+
+exports.MoneySend = MoneySend